Pull toplevel e.v.a.l.Root apart into subclasses
The file was getting long in the teeth but, more importantly, the inner
classes need to be pulled apart because spark3.0 has a completely
different API and will need to be refactored anyway
